Add guest users for app

+++ functions/src/index.ts
@@ -2,6 +2,7 @@
 import * as functions from "firebase-functions";
 import * as admin from "firebase-admin";
 import * as corsFactory from "cors";
+import * as nodemailer from "nodemailer";
 
 admin.initializeApp();
 
@@ -27,6 +28,12 @@ export const deleteUnverifiedUsers =
       const uid = doc.id;
       const userRecord = await admin.auth().getUser(uid);
 
+      // Skip if it's an anonymous account
+      if (userRecord.providerData.length === 0) {
+        return;
+      }
+
+      // Only delete if it's not an anonymous account and the email is not verified
       if (!userRecord.emailVerified) {
         await admin.auth().deleteUser(uid);
         await doc.ref.delete();
@@ -39,36 +46,120 @@ export const deleteUnverifiedUsers =
     functions.logger.info(`Processed ${usersSnapshot.size} users.`);
   });
 
+export const deleteAnonymousUsers = functions.pubsub.schedule("every 24 hours").onRun(async () => {
+    const currentTime = admin.firestore.Timestamp.now();
+    const cutoffTime = currentTime.toMillis() - (1 * 1 * 60 * 60 * 1000); // 31 days in milliseconds
+    const cutoffTimestamp = admin.firestore.Timestamp.fromMillis(cutoffTime);
 
-export const handleUserCreate = functions.auth.user().onCreate(async (user) => {
-    const isGoogleSignIn = user.providerData
-        .some((provider) => provider.providerId === "google.com");
-    const newUser = {
-        uid: user.uid,
-        displayName: user.displayName,
-        email: user.email,
-        emailVerified: isGoogleSignIn || user.emailVerified,
-        registrationTimestamp: admin.firestore.FieldValue.serverTimestamp(),
-        dreamTokens: 10, // Add the dreamTokens field with an initial value of 10
-        // Add any other fields you need
-    };
-    await firestore.collection("users").doc(user.uid).set(newUser);
-    functions.logger
-        .info(`Created new user with UID: ${user.uid} using ${isGoogleSignIn ? "Google Sign In" : "email and password"}.`);
+    const usersSnapshot = await firestore
+        .collection("users")
+        .where("registrationTimestamp", "<=", cutoffTimestamp)
+        .get();
+
+    const deletePromises = usersSnapshot.docs.map(async (doc) => {
+        const uid = doc.id;
+        const userRecord = await admin.auth().getUser(uid);
+
+        // Only delete if it's an anonymous account
+        if (userRecord.providerData.length === 0) {
+            await admin.auth().deleteUser(uid);
+            await doc.ref.delete();
+            functions.logger.info(`Deleted anonymous user with UID: ${uid}.`);
+        }
+    });
+
+    await Promise.all(deletePromises);
+
+    functions.logger.info(`Processed ${usersSnapshot.size} users.`);
 });
 
-exports.handleEmailVerificationUpdate = functions.auth.user().onUpdate(async (userRecord, context) => {
-    const { uid } = userRecord;
-    const userEmailVerified = userRecord.emailVerified;
+export const handleAccountLinking = functions.https.onCall(async (data, context) => {
+    const {permanentUid, anonymousUid} = data;
 
-    const userDoc = admin.firestore().collection("users").doc(uid);
-    const userDocSnapshot = await userDoc.get();
-    const userData = userDocSnapshot.data();
+    functions.logger.info("Starting account linking. Anon UID: " + anonymousUid + ", Perm UID: " + permanentUid);
 
-    if (userData && !userData.emailVerified && userEmailVerified) {
-        functions.logger.info(`Updated emailVerified for UID: ${uid} to true.`);
+    if (!context.auth || !context.auth.uid) {
+        functions.logger.error("No context.auth or context.auth.uid found. The function must be called while authenticated.");
+        throw new functions.https.HttpsError("unauthenticated", "The function must be called while authenticated.");
+    }
 
-        await userDoc.update({ emailVerified: true });
+    if (context.auth.uid !== permanentUid) {
+        functions.logger.error("Context.auth.uid (" + context.auth.uid +
+        ") does not match permanent UID (" + permanentUid + "). The authenticated user does not have sufficient permissions.");
+        throw new functions.https.HttpsError("permission-denied", "The authenticated user does not have sufficient permissions.");
+    }
+
+    const anonymousDreamsRef = firestore.collection("users").doc(anonymousUid).collection("my_dreams");
+    const permanentDreamsRef = firestore.collection("users").doc(permanentUid).collection("my_dreams");
+
+    let lastSnapshot = null;
+
+    do {
+        const anonymousDreamsSnapshot = await (lastSnapshot ?
+            anonymousDreamsRef.startAfter(lastSnapshot.docs[lastSnapshot.docs.length - 1]).limit(500).get() :
+            anonymousDreamsRef.limit(500).get());
+
+        const batch = firestore.batch();
+
+        // Log the number of dreams
+        functions.logger.info("Anon UID " + anonymousUid + " has " + anonymousDreamsSnapshot.size + " dreams.");
+
+        // Create new dream docs in the permanent account and keep in the anonymous account
+        anonymousDreamsSnapshot.forEach((doc) => {
+            const newDreamRef = permanentDreamsRef.doc(doc.id);
+            functions.logger.info("Transferring dream with ID " + doc.id + " to permanent account (" + permanentUid + ").");
+            batch.set(newDreamRef, doc.data());
+        });
+
+        // Commit the batch
+        await batch.commit();
+
+        functions.logger.info("Successfully committed batch");
+
+        lastSnapshot = anonymousDreamsSnapshot;
+    } while (lastSnapshot && lastSnapshot.size === 500);
+
+    // Fetch the user record for the permanent user
+    const permanentUserRecord = await admin.auth().getUser(permanentUid);
+
+    // If the permanent user's email is verified, delete the anonymous account
+    if (permanentUserRecord.emailVerified) {
+        await admin.auth().deleteUser(anonymousUid);
+        await firestore.collection("users").doc(anonymousUid).delete();
+        functions.logger.info(`Deleted anonymous user with UID: ${anonymousUid} after transfer.`);
+    }
+
+    return {success: true, message: "All dreams transferred successfully."};
+});
+
+
+export const handleUserCreate = functions.auth.user().onCreate(async (user) => {
+    // Check if it's an anonymous account
+    const isAnonymous = user.providerData.length === 0;
+
+    // Check if the user signed in with Google
+    const isGoogleSignIn = user.providerData
+        .some((provider) => provider.providerId === "google.com");
+
+    // Prepare the new user document
+    const newUser = {
+        uid: user.uid,
+        displayName: user.displayName || "Anonymous",
+        email: user.email || "",
+        emailVerified: isGoogleSignIn || user.emailVerified || false,
+        registrationTimestamp: admin.firestore.FieldValue.serverTimestamp(),
+        dreamTokens: isAnonymous ? 0 : 50, // No dreamTokens for anonymous users
+        // Add any other fields you need
+    };
+
+    // Write the user document to Firestore
+    await firestore.collection("users").doc(user.uid).set(newUser);
+
+    // Log the creation
+    if (isAnonymous) {
+        functions.logger.info(`Created new anonymous user with UID: ${user.uid}.`);
+    } else {
+        functions.logger.info(`Created new user with UID: ${user.uid} using ${isGoogleSignIn ? "Google Sign In" : "email and password"}.`);
     }
 });
 
@@ -95,3 +186,63 @@ exports.handlePurchaseVerification = functions.https.onRequest(async (req, res) 
     }
   });
 });
+
+export const createAccount = functions.https.onCall(async (data, context) => {
+    // Get the user's email and password from the data passed to the function
+    const userEmail = data.email;
+    const userPassword = data.password;
+
+    // Create the user with the Firebase Admin SDK
+    const userRecord = await admin.auth().createUser({
+        email: userEmail,
+        password: userPassword,
+        emailVerified: false
+    });
+
+    // Generate the verification email link
+    const verificationLink = await admin.auth().generateEmailVerificationLink(userEmail);
+
+    // Get the application's email and password from Firebase environment variables
+    const appEmail = functions.config().email.credentials;
+    const appPassword = functions.config().password.credentials;
+
+    // Get the OAuth2 client ID and client secret from Firebase environment variables
+    const clientID = functions.config().oauth.client_id;
+    const clientSecret = functions.config().oauth.client_secret;
+
+    const oauth2Client = new google.auth.OAuth2(
+        clientID, // Client ID
+        clientSecret, // Client Secret
+        'https://developers.google.com/oauthplayground' // Redirect URL
+    );
+
+    // Replace the following line with the method to get the refresh token
+    // const refreshToken = 'YOUR_REFRESH_TOKEN';
+
+    const { token } = await oauth2Client.getAccessToken();
+
+    // Set up nodemailer with your SMTP details
+    const transporter = nodemailer.createTransport({
+        service: "gmail",
+        auth: {
+            type: 'OAuth2',
+            user: appEmail,
+            clientId: clientID,
+            clientSecret: clientSecret,
+            refreshToken: refreshToken,
+            accessToken: token
+        }
+    });
+
+    // Send the verification email
+    await transporter.sendMail({
+        from: appEmail,
+        to: userEmail,
+        subject: "Email Verification",
+        text: `Please verify your email by clicking on the following link: ${verificationLink}`,
+        html: `<p>Please verify your email by clicking on the following link: <a href="${verificationLink}">${verificationLink}</a></p>`
+    });
+
+    return { uid: userRecord.uid };
+});
+
